refactor: move cs.clamscanner to com.zimbra.clam (original namespace)
