+++ store/src/main/java/com/zimbra/cs/extension/ExtensionUtil.java
@@ -9,7 +9,7 @@ import com.zimbra.common.localconfig.LC;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.util.Log.Level;
 import com.zimbra.common.util.ZimbraLog;
-import com.zimbra.cs.clamscanner.ClamScannerExtension;
+import com.zimbra.clam.ClamScannerExtension;
 import com.zimbra.cs.ephemeral.EphemeralStore;
 import com.zimbra.cs.ephemeral.EphemeralStore.Factory;
 import com.zimbra.cs.nginx.NginxLookupExtension;
